refactor: Rewrite package with utilites
