+++ server/utilites/common/funcs.py
@@ -1,0 +1,5 @@
+from ulid import new, ULID
+
+
+def make_ULID() -> ULID:
+    return new()
